BibInternal.AbstractEntry — Type

abstract type AbstractEntry end

Abstract Entry supertype.

BibInternal.Fields — Type

Fields = Dict{String, String}

Stores the fields (name => value) of an entry.

BibInternal.Access — Type

struct Access
    doi::String
    howpublished::String
    url::String
end

Store the online access of an entry as a String. Handles the fields doi and url and the arXiV entries. For additional fields or entries, please fill an issue or make a pull request.

BibInternal.Access — Method

Access(fields::Fields)

Construct the online access information based on the entry fields.

BibInternal.Date — Type

struct Date
    day::String
    month::String
    year::String
end

Store the date information as day, month, and year.

BibInternal.Date — Method

Date(fields::Fields)

Construct the date information based on the entry fields.

BibInternal.Entry — Type

struct Entry <: AbstractEntry
    access::Access
    authors::Names
    booktitle::String
    date::Date
    editors::Names
    eprint::Eprint
    id::String
    in::In
    fields::Dict{String,String}
    title::String
    type::String
end

Generic Entry type. If some construction rules are required, it should be done beforehand. Check bibtex.jl as the example of rules implementation for BibTeX format.

BibInternal.Entry — Method

Entry(id::String, fields::Fields)

Construct an entry with a unique id and a list of Fields.

BibInternal.Eprint — Type

struct Eprint
    archive_prefix::String
    eprint::String
    primary_class::String
end

Store the information related to arXiv eprint format.

BibInternal.Eprint — Method

Eprint(fields::Fields)

Construct the eprint arXiv information based on the entry fields. Handle old and current arXiv format.

BibInternal.In — Type

struct In
    address::String
    chapter::String
    edition::String
    institution::String
    journal::String
    number::String
    organization::String
    pages::String
    publisher::String
    school::String
    series::String
    volume::String
end

Store all the information related to how an entry was published.

BibInternal.In — Method

In(fields::Fields)

Construct the information of how an entry was published based on its fields

BibInternal.Name — Method

Name(str::String)

Decompose without ambiguities a name as particle (optional) last, junior (optional), first middle (optional) based on BibTeX possible input. As for BibTeX, the decomposition of a name in the form of first last is also possible, but ambiguities can occur.

BibInternal.names — Method

names(str::String)

Decompose into parts a list of names in BibTeX compatible format. That is names sparated by and.