BibInternal.entries — Constant

const entries = [
    :article,
    :book,
    :booklet,
    :inbook,
    :incollection,
    :inproceedings,
    :manual,
    :mastersthesis,
    :misc,
    :phdthesis,
    :proceedings,
    :techreport,
    :unpublished,
]

List of possible entries (currently based on bibtex). Keep it sorted for readability.

BibInternal.fields — Constant

const fields = [
    :address,
    :annote,
    :archivePrefix,
    :author,
    :booktitle,
    :chapter,
    :crossref,
    :edition,
    :editor,
    :eprint,
    :howpublished,
    :institution,
    :isbn,
    :issn,
    :journal,
    :key,
    :month,
    :note,
    :number,
    :organization,
    :pages,
    :primaryClass,
    :publisher,
    :school,
    :series,
    :title,
    :type,
    :volume,
    :year
]

List of possible fields (currently based on bibtex). Keep it sorted for readability

BibInternal.maxfieldlength — Constant

const maxfieldlength

For output formatting purpose, for instance, export to BibTeX format.

BibInternal.Fields — Type

Fields = Dict{String, String}. Stores the fields name => value of an entry.

BibInternal.arxive_url — Method

arxive_url(fields::Fields)

Make an arxiv url from an eprint entry. Work with both old and current arxiv BibTeX format.

BibInternal.erase_spaces — Method

erase_spaces(str::String)

Erase extra spaces, i.e. r"[ ]+", from str and return a new string.

BibInternal.get_delete! — Method

get_delete!(fields::Fields, key::String)

Get the value of a field and delete it afterward.

BibInternal.space — Method

space(field::Symbol)

Return the amount of spaces needed to export entries, for instance to BibTeX format.